增加登录逻辑

+++ app/src/main/java/com/ddx/compose/model/User.kt
@@ -2,7 +2,7 @@ package com.ddx.compose.model
 
 import java.io.Serializable
 
-data class User(var username: String, var status: UserStatus, var expireTime: Int) :
+data class User(var username: String, var status: UserStatus, var expireTime: Long) :
     Serializable
 
 enum class UserStatus{
